requirements.txt
altair @ file:///croot/altair_1743016723624/work
attrs @ file:///croot/attrs_1734533101012/work
blinker @ file:///croot/blinker_1737448726027/work
Brotli @ file:///home/conda/feedstock_root/build_artifacts/brotli-split_1666788425425/work
cachetools @ file:///croot/cachetools_1738224668060/work
certifi @ file:///croot/certifi_1754570635119/work/certifi
cffi @ file:///home/conda/feedstock_root/build_artifacts/cffi_1725571112467/work
charset-normalizer @ file:///home/conda/feedstock_root/build_artifacts/charset-normalizer_1754767332901/work
click @ file:///croot/click_1744271578095/work
contourpy @ file:///home/conda/feedstock_root/build_artifacts/contourpy_1727293517607/work
cycler @ file:///home/conda/feedstock_root/build_artifacts/cycler_1733332471406/work
fonttools @ file:///home/conda/feedstock_root/build_artifacts/fonttools_1755223865520/work
gitdb @ file:///croot/gitdb_1753362118172/work
GitPython @ file:///croot/gitpython_1720455034730/work
h2 @ file:///home/conda/feedstock_root/build_artifacts/h2_1738578511449/work
hpack @ file:///home/conda/feedstock_root/build_artifacts/hpack_1737618293087/work
hyperframe @ file:///home/conda/feedstock_root/build_artifacts/hyperframe_1737618333194/work
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1733211830134/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1736252299705/work
Jinja2 @ file:///croot/jinja2_1741710844255/work
joblib @ file:///home/conda/feedstock_root/build_artifacts/joblib_1748019130050/work
jsonschema @ file:///croot/jsonschema_1753206402902/work
jsonschema-specifications @ file:///croot/jsonschema-specifications_1699032386549/work
kiwisolver @ file:///home/conda/feedstock_root/build_artifacts/kiwisolver_1725459266648/work
MarkupSafe @ file:///croot/markupsafe_1738584038848/work
matplotlib==3.9.4
munkres==1.1.4
narwhals @ file:///croot/narwhals_1742845957875/work
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1668919081525/work
packaging @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_packaging_1745345660/work
pandas @ file:///home/conda/feedstock_root/build_artifacts/pandas_1752081702369/work
patsy @ file:///home/conda/feedstock_root/build_artifacts/patsy_1733792384640/work
pillow @ file:///croot/pillow_1752524526621/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_platformdirs_1746710438/work
pooch @ file:///home/conda/feedstock_root/build_artifacts/pooch_1754941678315/work
protobuf==4.25.3
pyarrow==17.0.0
pycparser @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pycparser_1733195786/work
pyparsing @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_pyparsing_1753873557/work
PySide6==6.7.3
PySocks @ file:///home/conda/feedstock_root/build_artifacts/pysocks_1733217236728/work
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_python-dateutil_1751104122/work
pytz @ file:///home/conda/feedstock_root/build_artifacts/pytz_1742920838005/work
PyYAML @ file:///croot/pyyaml_1728657952215/work
referencing @ file:///croot/referencing_1699012038513/work
requests @ file:///home/conda/feedstock_root/build_artifacts/requests_1755614211359/work
rpds-py @ file:///croot/rpds-py_1736541261634/work
scikit-learn @ file:///home/conda/feedstock_root/build_artifacts/scikit-learn_1685023709438/work
scipy @ file:///home/conda/feedstock_root/build_artifacts/scipy-split_1683900212083/work/base/dist/scipy-1.10.1-cp39-cp39-linux_x86_64.whl#sha256=9cf5ab9346500556bd9381db5c987124e630ef8038f1acbc0a9f162bc143903d
seaborn @ file:///home/conda/feedstock_root/build_artifacts/seaborn-split_1733730015268/work
shiboken6==6.7.3
six @ file:///home/conda/feedstock_root/build_artifacts/bld/rattler-build_six_1753199211/work
smmap @ file:///croot/smmap_1752748938695/work
statsmodels @ file:///home/conda/feedstock_root/build_artifacts/statsmodels_1751917446772/work
streamlit @ file:///home/task_175733971257161/conda-bld/streamlit-split_1757340213987/work/streamlit
tenacity @ file:///croot/tenacity_1730302500303/work
threadpoolctl @ file:///home/conda/feedstock_root/build_artifacts/threadpoolctl_1741878222898/work
toml @ file:///tmp/build/80754af9/toml_1616166611790/work
torch==1.13.1
torchaudio==0.13.1
torchvision==0.14.1
tornado @ file:///home/conda/feedstock_root/build_artifacts/tornado_1754731992698/work
typing_extensions @ file:///croot/typing_extensions_1756280817316/work
tzdata @ file:///home/conda/feedstock_root/build_artifacts/python-tzdata_1742745135198/work
unicodedata2 @ file:///home/conda/feedstock_root/build_artifacts/unicodedata2_1736692503055/work
urllib3 @ file:///home/conda/feedstock_root/build_artifacts/urllib3_1750271362675/work
watchdog @ file:///croot/watchdog_1740409902579/work
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1749421620841/work
zstandard==0.23.0
